+++ src/app/desktopLayout/page.js
@@ -1,0 +1,13 @@
+import Home from "./components/home"
+import Services from "./components/services"
+import AboutUs from "./components/aboutUs"
+
+export default function DesktopLayout(){
+    return(
+        <>
+        <Home />
+        <Services />
+        <AboutUs />
+        </>
+    )
+}
