+++ src/app/mobileLayout/components/mobileServices.js
@@ -4,7 +4,8 @@ export default function MobileServices(){
     return(
         <div className="flex justify-center items-center w-full">
             <div className="
-                grid gap-[0.5vw] md:mt-8 
+                md:hidden
+                grid gap-[0.5vw]
                 grid-cols-[repeat(31,_minmax(0,_2vw))]
                 grid-rows-[repeat(45,_minmax(0,_2vw))]
             ">
