+++ src/app/mobileLayout/page.js
@@ -1,15 +1,15 @@
+import NavBar from "../modules/NavBar"
 import MobileHome from "./components/mobileHome"
 import MobileServices from "./components/mobileServices"
 import MobileAboutUs from "./components/mobileAboutUs"
 
-const linkContattaci="../contattaci"
-
 export default function MobileLayout(){
     return(
-        <div className="w-screen h-screen">
+        <>
+        <NavBar />
         <MobileHome />
         <MobileServices />
         <MobileAboutUs />
-        </div>
+        </>
     )
 }
