+++ src/app/page.js
@@ -1,7 +1,11 @@
 import MobileLayout from "./mobileLayout/page";
+import DesktopLayout from "./desktopLayout/page";
 
 export default function SplashScreen() {
   return (
-    <MobileLayout />
+    <div className="w-screen h-screen">
+      <MobileLayout />
+      <DesktopLayout />
+    </div>
   );
 }
